Fix clang compiler and linker errors, fix map drawing

+++ src/opengl/Shape.hh
@@ -21,11 +21,11 @@ class Shape {
   template <typename T>
   void gen(uint32_t& vb, std::vector<T>& list);
   template <typename T>
-  void gen(uint32_t& vb, std::vector<T>& list, uint32_t stride, uint32_t shaderParam, uint32_t numElements, uint32_t offset);
+  void gen(uint32_t& vb, std::vector<T>& list, uint32_t stride, uint32_t shaderParam, uint32_t numElements, uint64_t offset);
   template <typename T>
   void gen(uint32_t& vb, std::vector<T>& list, uint32_t stride,
-           uint32_t shaderParam1, uint32_t numElements1, uint32_t offset1,
-           uint32_t shaderParam2, uint32_t numElements2, uint32_t offset2);
+           uint32_t shaderParam1, uint32_t numElements1, uint64_t offset1,
+           uint32_t shaderParam2, uint32_t numElements2, uint64_t offset2);
  public:
   Shape(Canvas* parent) : parentCanvas(parent) {}
   virtual ~Shape();
