+++ src/opengl/Shape_impl.hh
@@ -19,7 +19,7 @@ inline void Shape::gen(uint32_t& vb, std::vector<T>& list) {
 
 // generic function to generate a vertex buffer and set up a single vertex attrib pointer
 template <typename T>
-inline void Shape::gen(uint32_t& vb, std::vector<T>& list, uint32_t stride,   uint32_t shaderParam, uint32_t numElements, uint32_t offset) {
+inline void Shape::gen(uint32_t& vb, std::vector<T>& list, uint32_t stride,   uint32_t shaderParam, uint32_t numElements, uint64_t offset) {
   get(vb, list);
 	glVertexAttribPointer(shaderParam, numElements, GL_FLOAT, GL_FALSE, stride, (void*)offset);
 }
@@ -28,8 +28,8 @@ inline void Shape::gen(uint32_t& vb, std::vector<T>& list, uint32_t stride,   ui
 // generic function to generate a vertex buffer and set up two vertex attrib pointers
 template <typename T>
 inline void Shape::gen(uint32_t& vb, std::vector<T>& list, uint32_t stride,
-                       uint32_t shaderParam1, uint32_t numElements1,  uint32_t offset1,
-                       uint32_t shaderParam2, uint32_t numElements2, uint32_t offset2) {
+                       uint32_t shaderParam1, uint32_t numElements1,  uint64_t offset1,
+                       uint32_t shaderParam2, uint32_t numElements2, uint64_t offset2) {
   get(vb, list);
 	glVertexAttribPointer(shaderParam1, numElements1, GL_FLOAT, GL_FALSE, stride, (void*)offset1);
 	glVertexAttribPointer(shaderParam2, numElements2, GL_FLOAT, GL_FALSE, stride, (void*)offset2);
